STOP: Support Suspend Entry/Exit and Fix Pig Collision
1) also cleanup todos in Stop Hcode
2) make STOP3 complete trans in SSH

Key_Cronus_Test=PM_REGRESS

Original-Change-Id: I28a146e15e455f09f8d8ff588e122d5ecf34110a
CQ: SW416550
CQ: HW437955
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/54660
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Cronus HW CI <cronushw-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>

+++ import/chips/p9/common/pmlib/include/wof_sgpe_pgpe_api.h
@@ -99,12 +99,12 @@ typedef union
     uint64_t value;
     struct
     {
-        uint32_t msg_num                : 4;
-        uint32_t update_type            : 1;
-        uint32_t reserved               : 3;
-        uint32_t return_code            : 8;
-        uint32_t active_cores           : 24;
-        uint32_t return_active_cores    : 24;
+        uint64_t msg_num                : 4;
+        uint64_t update_type            : 1;
+        uint64_t reserved               : 3;
+        uint64_t return_code            : 8;
+        uint64_t active_cores           : 24;
+        uint64_t return_active_cores    : 24;
     } fields;
 } ipcmsg_s2p_update_active_cores_t;
 
@@ -113,16 +113,16 @@ typedef union
     uint64_t value;
     struct
     {
-        uint32_t msg_num                : 4;
-        uint32_t update_type            : 1;
-        uint32_t entry_type             : 1;
-        uint32_t reserved               : 2;
-        uint32_t return_code            : 8;
-        uint32_t requested_quads        : 6;
-        uint32_t reserved0              : 2;
-        uint32_t return_active_quads    : 6;
-        uint32_t reserved1              : 2;
-        uint32_t pad                    : 32;
+        uint64_t msg_num                : 4;
+        uint64_t update_type            : 1;
+        uint64_t entry_type             : 1;
+        uint64_t reserved               : 2;
+        uint64_t return_code            : 8;
+        uint64_t requested_quads        : 6;
+        uint64_t reserved0              : 2;
+        uint64_t return_active_quads    : 6;
+        uint64_t reserved1              : 2;
+        uint64_t pad                    : 32;
     } fields;
 } ipcmsg_s2p_update_active_quads_t;
